Migrate some date expressions to operators
